+++ src/test/kotlin/com/pestphp/pest/tests/types/fixtures/thisField/beforeEachCompletion.php
@@ -1,0 +1,9 @@
+<?php
+
+beforeEach(function () {
+    $this->foo = 1;
+});
+
+it('has home', function () {
+    $this-><caret>
+});